[Agent] unify event dispatch helpers into service
